[refactor] mate 패키지 이름을 mates로 변경
