+++ src/main/java/klieme/artdiary/mates/data_access/repository/MateRepository.java
@@ -1,4 +1,4 @@
-package klieme.artdiary.mate.data_access.repository;
+package klieme.artdiary.mates.data_access.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import klieme.artdiary.mate.data_access.entity.MateEntity;
+import klieme.artdiary.mates.data_access.entity.MateEntity;
 
 @Repository
 public interface MateRepository extends JpaRepository<MateEntity, Long> {
